requirements-dev.txt
wheel==0.36.2
pyspark==3.1.1
flake8==3.9.0
mypy==0.812
pytest==6.2.3
black==21.6b0
pre-commit==2.13.0
pandas==1.2.3
apache-airflow==2.0.2
apache-airflow-providers-apache-livy==1.1.0
